rename AdminConstants => constants
